+++ src/refactoring/components/index.ts
@@ -1,3 +1,2 @@
 export * from './PageLayout';
 export * from './Navigation';
-export * from './ProductCard';
