+++ src/refactoring/pages/cart/components/index.ts
@@ -1,1 +1,3 @@
 export * from './ProductList';
+export * from './CartList';
+export * from './ProductCard';
